move tests to test_body_map
